cortexSysTodo/app/jwt.py
# app/jwt.py
from datetime import datetime, timedelta
from jose import JWTError, jwt
import secrets
from .config import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES



def create_access_token(user_id: int):
    """ساخت JWT Token با زمان انقضا"""

    to_encode = {"sub": str(user_id)}
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

def decode_access_token(token: str):
    """Decode و برگردوندن payload"""

    try:
            payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
            return payload
    except JWTError:
            return None
